// src/typing/enums.ts
export enum MaskEnum {
    CPF = "###.###.###-##",
    CNPJ = "##.###.###/####-##"
}

export enum TransactionPeriodEnum {
    NONE = 'NONE',
    MONTHLY = 'MONTHLY',
    WEEKLY = 'WEEKLY'
}

export enum WeekdaysEnum {
    SUNDAY = 'SUNDAY',
    MONDAY = 'MONDAY',
    TUESDAY = 'TUESDAY',
    WEDNESDAY = 'WEDNESDAY',
    THUSRDAY = 'THURSDAY',
    FRIDAY = 'FRIDAY',
    SATURDAY = 'SATURDAY'
}